refactor: ForecastStatusについて再度整理
