+++ src/app/(app)/home/page.tsx
@@ -3,7 +3,7 @@ import { Stack, Text, VStack } from '@chakra-ui/react';
 import PageTitle from '@/components/PageTitle';
 import TitleWithIcon from '@/components/TitleWithIcon';
 import { Lightbulb, Onigiri } from '@phosphor-icons/react';
-import ForecastStatusSection from '@/features/forecast-status/components/ForecastStatusSection';
+import ForecastStatusSection from '@/features/forecast-status/ForecastStatusSection';
 import AlertContainer from '@/features/hint/components/AlertContainer';
 
 const Page = () => {
