+++ src/features/forecast-status/ForecastStatusSection.tsx
@@ -1,4 +1,4 @@
-import { Pending } from '@/features/forecast-status/PendingStatus';
+import { Pending } from '@/features/forecast-status/components/Pending';
 
 const ForecastStatusSection = () => {
   return <Pending current={4} max={10} />;
